Move getInnerRegion to CSGUtils file
