+++ modules/reactor/include/meshgenerators/PinMeshGenerator.h
@@ -44,18 +44,6 @@ protected:
   std::vector<std::reference_wrapper<const CSG::CSGSurface>>
   getOuterRadialSurfaces(unsigned int radial_index, Real halfpitch, CSG::CSGBase & csg_obj);
 
-  /**
-   * Get inner region of given surfaces, defined as the intersection
-   * of halfspaces of each surface. Here, the halfspace direction is determined based on the origin.
-   * @param radial_surfaces List of references to surfaces used to define inner region. This should
-   * be defined with the minimum number of surfaces to enclose the region
-   * @param origin Point used to determine halfspace direction when defining intersected region
-   * @return inner region defined by surfaces
-   */
-  CSG::CSGRegion
-  getInnerRegion(const std::vector<std::reference_wrapper<const CSG::CSGSurface>> & radial_surfaces,
-                 const libMesh::Point & origin = Point(0, 0, 0));
-
   /// The id number for this pin type
   const subdomain_id_type _pin_type;
